std::min/maxを使う

+++ BonCtrl/SendUDP.cpp
@@ -79,7 +79,7 @@ void CSendUDP::SendData(BYTE* pbBuff, DWORD dwSize)
 {
 	for( DWORD dwRead=0; dwRead<dwSize; ){
 		//ペイロード分割。BonDriver_UDPに送る場合は受信サイズ48128以下でなければならない
-		int iSendSize = min(max(m_uiSendSize, 188), dwSize - dwRead);
+		int iSendSize = min(max((int)m_uiSendSize, 188), (int)(dwSize - dwRead));
 		for( size_t i=0; i<SockList.size(); i++ ){
 			int iRet = sendto(SockList[i].sock, (char*)(pbBuff + dwRead), iSendSize, 0, (struct sockaddr *)&SockList[i].addr, (int)SockList[i].addrlen);
 			if( iRet == SOCKET_ERROR ){
